Move shared types to dedicated file
